+++ cmd/tags_cmd.go
@@ -26,8 +26,8 @@ import (
 )
 
 type TagsCmd struct {
-	AccountId   int64  `kong:"optional,name='account',short='A',default=-1,help='Filter regsults based on AWS AccountID',env='AWS_SSO_ACCOUNTID'"`
-	Role        string `kong:"optional,name='role',short='R',help='Filter results based on AWS Role Name',env='AWS_SSO_ROLE'"`
+	AccountId   int64  `kong:"optional,name='account',short='A',help='Filter regsults based on AWS AccountID'"`
+	Role        string `kong:"optional,name='role',short='R',help='Filter results based on AWS Role Name'"`
 	ForceUpdate bool   `kong:"optional,name='force-update',help='Force account/role cache update'"`
 }
 
@@ -62,14 +62,20 @@ func (cc *TagsCmd) Run(ctx *RunContext) error {
 	roles := []*sso.AWSRoleFlat{}
 
 	// If user has specified an account (or account + role) then limit
-	if ctx.Cli.Tags.AccountId != -1 {
+	if ctx.Cli.Tags.AccountId != 0 {
 		for _, fRole := range set.Cache.Roles.GetAccountRoles(ctx.Cli.Tags.AccountId) {
-			if ctx.Cli.Tags.Role != "" {
+			if ctx.Cli.Tags.Role == "" {
 				roles = append(roles, fRole)
 			} else {
 				if fRole.RoleName == ctx.Cli.Tags.Role {
 					roles = append(roles, fRole)
 				}
+			}
+		}
+	} else if ctx.Cli.Tags.Role != "" {
+		for _, v := range set.Cache.Roles.GetAllRoles() {
+			if v.RoleName == ctx.Cli.Tags.Role {
+				roles = append(roles, v)
 			}
 		}
 	} else {